+++ libs/+state/src/lib/state/actions/index.ts
@@ -1,3 +1,2 @@
-export * from './arrivals.action'
-export * from './stations.action'
-export * from './user.action'
+export * from './api.action'
+export * from './local.action'
